% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_and_mean.R
\name{plot_distribution_and_mean}
\alias{plot_distribution_and_mean}
\title{Plot the Histogram of Numerical Variable and Bin Adjustment}
\usage{
plot_distribution_and_mean(
  dataframe,
  num_variable,
  bin_number = 30,
  decimals = 3
)
}
\arguments{
\item{dataframe}{any dataframe that has at least one numerical variable. The param name 'dataframe' is used to make it clear that the input must be a dataframe, not a vector.}

\item{num_variable}{any numerical variable in the dataframe that the user wants to plot. The param name 'num_variable' demonstrates that the input should be a numeric variable, though there is a if statement that does check for this.}

\item{bin_number}{the number of bins the user wants to assign that works best for the distribution of the variable. This must be a positive integer greater than 0. The param name 'bin_number' is used to ensure that that it is clear that the input for this variable is a number of bins. A default is set to 30 as this is the default number of bins set by ggplot geom_histogram().}

\item{decimals}{the number of digits to round the mean value to in the output. The default is set to 3 digits (decimals). Must be a positive integer.}
}
\value{
the histogram plot (plot_data) and the mean value for the variable, outputted in a sentence
("The mean of \if{html}{\out{<variable>}} is \if{html}{\out{<mean value>}}"). The list is used so that the outputs are assigned as components.
}
\description{
This purpose of this function is to construct a histogram based on a numerical variable that the user
inputs from a specific dataframe, and adjust the number of bins as necessary to make the graph
readable. A x-intercept line is added to show the user the mean value for the numerical variable based
on the histogram distribution, and the mean value of the variable is outputted
as a sentence along with the plot.
}
\examples{
plot_distribution_and_mean(mtcars, wt, 10, 4)
plot_distribution_and_mean(mtcars, mpg)
}
